+++ client/components/NavBar/BottomNavBar/BottomNavBar.scss
