+++ package.json
@@ -27,11 +27,8 @@
     "@iconify/icons-dashicons": "^1.0.4",
     "@iconify/icons-mdi": "^1.0.79",
     "@iconify/react": "^1.1.3",
-    "@zeit/next-css": "^1.0.1",
     "@zeit/next-sass": "^1.0.1",
     "bcryptjs": "^2.4.3",
-    "bootstrap": "^4.4.1",
-    "bootstrap-css-only": "^4.4.1",
     "dotenv": "^8.2.0",
     "isomorphic-unfetch": "^3.0.0",
     "js-cookie": "^2.2.1",
