+++ pages/index.jsx
@@ -3,7 +3,7 @@ import React from "react";
 const HomePage = () => {
   return (
     <>
-      <h2>Welcome to Next.js!</h2>
+      <h2>Liv2BGirl</h2>
     </>
   );
 };
