+++ public/static/colors.scss
@@ -1,0 +1,5 @@
+// Global Colors
+$background: #f2f2f2;
+$light-gray: #c4c4c4;
+$gray: #828282;
+$text: #4f4f4f;
